(fix): Incremental Parenting Fix
This is a proposed fix for an issue observed where the incremental backups aren't being parented to the most recent full backup.  In my environment, I retain 60 daily backups, and am consistently seeing the hourly backup parenting to the oldest retained full backup.

This change introduces a reverse sort based on the `created_at` attribute.  After applying this change, my backups are consistently parenting to the correct, latest full backup.